payment page draft

+++ app/controllers/orders_controller.rb
@@ -1,34 +1,24 @@
 class OrdersController < ApplicationController
   def add_product_quantity_quick_shop
-    # create/get cart
     order_pending = current_user.pending_order_in_cart
-    # associate product and qty to cart
     product = Product.find(params[:product_id])
-    # check if order_product exist
     order_product = order_pending.orders_products.where(product: product).first
-    # if exists, increment qty by 1
     if order_product
       order_product.quantity += 1
       order_product.price = product.price
       order_product.save
-    # if not, add product to the cart w/ qty = 1
     else
       OrdersProduct.create(quantity: 1, product: product, order: order_pending, price_cents: product.price_cents)
     end
-    # redirect to procreated_at).reverse.each do |orderduct show
     flash[:notice] = "You added one item to your shopping basket."
     authorize order_pending
     redirect_to category_products_path(product.category)
   end
 
   def remove_product_quantity_quick_shop
-    # create/get cart
     order_pending = current_user.pending_order_in_cart
-    # associate product and qty to cart
     product = Product.find(params[:product_id])
-    # check if order_product exist
     order_product = order_pending.orders_products.where(product: product).first
-    # if exists, decrease qty by 1
     if order_product
       if order_product.quantity == 0
         order_product.destroy
@@ -39,28 +29,21 @@ class OrdersController < ApplicationController
     end
     flash[:notice] = "You removed one item from your shopping basket."
     authorize order_pending
-    # redirect to product show
     redirect_to category_products_path(product.category)
   end
 
   def add_product_quantity
-    # create/get cart
     order_pending = current_user.pending_order_in_cart
-    # associate product and qty to cart
     product = Product.find(params[:product_id])
-    # check if order_product exist
     order_product = order_pending.orders_products.where(product: product).first
-    # if exists, increment qty by 1
     if order_product
       order_product.quantity += 1
       order_product.price = product.price
       order_product.save
-    # if not, add product to the cart w/ qty = 1
     else
       OrdersProduct.create(quantity: 1, product: product, order: order_pending, price_cents: product.price_cents)
     end
     flash[:notice] = "You added one item to your shopping basket."
-    # redirect to procreated_at).reverse.each do |orderduct show
     authorize order_pending
     redirect_to product_path(product)
   end
@@ -88,44 +71,30 @@ class OrdersController < ApplicationController
   end
 
   def add_product_quantity_checkout
-    # create/get cart
     order_pending = current_user.pending_order_in_cart
-    # associate product and qty to cart
     product = Product.find(params[:product_id])
-    # check if order_product exist
     order_product = order_pending.orders_products.where(product: product).first
-    # if exists, increment qty by 1
     if order_product
       order_product.quantity += 1
       order_product.price = product.price
       order_product.save
-    # if not, add product to the cart w/ qty = 1
     else
       OrdersProduct.create(quantity: 1, product: product, order: order_pending, price_cents: product.price_cents)
     end
-    # redirect to procreated_at).reverse.each do |orderduct show
     authorize order_pending
     redirect_to shopping_cart_orders_path(product)
   end
 
   def remove_product_quantity_checkout
-    # create/get cart
     order_pending = current_user.pending_order_in_cart
-    # associate product and qty to cart
     product = Product.find(params[:product_id])
-    # check if order_product exist
     order_product = order_pending.orders_products.where(product: product).first
-    # if exists, decrease qty by 1
     if order_product
-      if order_product.quantity == 0
-        order_product.destroy
-      else
-        order_product.quantity -= 1
-        order_product.save
-      end
+      order_product.quantity -= 1
+      order_product.save
+      order_product.destroy if order_product.quantity == 0
     end
     authorize order_pending
-    # redirect to product show
     redirect_to shopping_cart_orders_path(product)
   end
 
@@ -151,7 +120,7 @@ class OrdersController < ApplicationController
   end
 
   private
-  
+
   def calculate_total
     @results = []
     @order.orders_products.each do |product|
@@ -159,7 +128,6 @@ class OrdersController < ApplicationController
     end
     @results
   end
-
 end
   # def index
   #   @orders = policy_scope(Order)
